requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-google-genai
langchain-community
langchain-text-splitters
chromadb
pypdf